+++ src/main/java/com/example/task_management_system_ampada/services/UserServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.task_management_system_ampada.services;
 
+import com.example.task_management_system_ampada.exceptions.UserNotFoundException;
 import com.example.task_management_system_ampada.models.User;
 import com.example.task_management_system_ampada.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,12 +21,19 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Optional<User> findUserById(String id) {
-        return userRepository.findById(id);
+        if (userRepository.findById(id).isPresent())
+            return userRepository.findById(id);
+        else
+            throw new UserNotFoundException();
     }
 
     @Override
     public List<User> findAllUsers() {
-        return userRepository.findAll();
+        List<User> users = userRepository.findAll();
+        if (!users.isEmpty())
+            return users;
+        else
+            throw new UserNotFoundException();
     }
 
     @Override
@@ -40,12 +48,15 @@ public class UserServiceImpl implements UserService {
             user.setPassword(newUser.getPassword());
             return userRepository.save(user);
         }).orElseThrow(() -> {
-            throw new RuntimeException();
+            throw new UserNotFoundException();
         });
     }
 
     @Override
     public void deleteUserById(String id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id))
+            userRepository.deleteById(id);
+        else
+            throw new UserNotFoundException();
     }
 }
